+++ servant-server/src/Servant/Server.hs
@@ -81,7 +81,6 @@ import           Data.Proxy                         (Proxy)
 import           Network.Wai                        (Application)
 import           Servant.Server.Internal
 import           Servant.Server.Internal.Enter
-import           Servant.Server.Internal.ServantErr
 
 
 -- * Implementing Servers
